pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antidrone-radar-fdtd"
version = "0.2.0"
description = "Anti-drone radar pipeline (FMCW/CFAR/Tracking) + 3D FDTD (CPML + TF/SF + Huygens NTF) baseline"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
